Curate visit-vocab contract exports in their own module
index.ts stays under the line ceiling; new visit-vocab names still
have to be listed before they become public.
