+++ packages/contract/src/visit-vocab-contract.ts
@@ -1,0 +1,47 @@
+// Curated visit-stream vocabulary. Add names here to publish them from the contract.
+export {
+  ASSIST_STEPS,
+  BETA_WELCOME_STEPS,
+  CLI_ADAPTERS,
+  CLI_INSTALL_CHANNELS,
+  CLI_PLATFORM_OS,
+  CLI_STEPS,
+  CLI_VERIFY_STAGES,
+  CODE_COMPLETION_KINDS,
+  CODE_COMPLETION_OUTCOMES,
+  CODE_STEPS,
+  CREATE_STEPS,
+  EDITOR_STEPS,
+  HOW_TO_PLAY_VIAS,
+  INVITE_STEPS,
+  PLAY_VIAS,
+  REMIX_CONTROLS,
+  REMIX_PAINTED_VIAS,
+  REMIX_STEPS,
+  STUDIO_STEP_DETAILS,
+  STUDIO_STEPS,
+  VISIT_ROUTE_KINDS,
+  WAITLIST_STEPS,
+  type AssistStep,
+  type BetaWelcomeStep,
+  type CliAdapter,
+  type CliInstallChannel,
+  type CliPlatformOs,
+  type CliStep,
+  type CliVerifyStage,
+  type CodeCompletionKind,
+  type CodeCompletionOutcome,
+  type CodeStep,
+  type CreateStep,
+  type EditorStep,
+  type HowToPlayVia,
+  type InviteStep,
+  type PlayVia,
+  type RemixControl,
+  type RemixPaintedVia,
+  type RemixStep,
+  type StudioStep,
+  type StudioStepDetail,
+  type VisitRouteKind,
+  type WaitlistStep,
+} from './visit-vocab.js';
